Fixed the successful registration scenario
The first scenario wasn't clear and it was going into implementation
details. Scenarios are to be seen from the users point of view. The user
doesn't know whether there is a database or anything like that. Also,
the UI, although nice, is independent of the way the system works.
Scenarios ought not to reference UI (since it is going to be different
in the API, in Android, iOS and web) nor reference database since these
are just details that are to be decided upon later.

Finally, a little bit of understanding of Given/When/Then. Scenarios are
meant to be written from the user's perspective. That is why I is always
used. I do x or I get result y. And so on.

Given is to express things that are true before the process is started,
in this case it is true that the user hasn't registered yet, that the
user is in the registration form and that the user name he wants to use
(testuser) is available. When having several things to put either in the
Given, the When or the Then, the subsequent ones are prefixed with And
instead of repeating Given, When or Then.

When is to say what the user does. In this case, the user is filling all
the fields in the registration form. The best way is to do an
specification by example, in order to make absolutely clear what the
scenario is. So, the user sets its username to testuser, the password to
password, repeats the password, sets his email to testuser@gmail.com and
set his first name to Test and his last name to User.

Then is to tell what the result of the interaction is. In this case, the
user is taken into his newly created dashboard and a confirmation email
is sent to testuser@gmail.com