Errors with Then
Then is to denote expected outcomes. Think about Then as things that are
expected the system to do. When is things that the user do. These three
then statements are part of the successful registration scenario.

Also, don't put any statement that is part of an scenario outline after
the examples. The examples are always the last thing in an scenario
outline.