Fixing misunderstanding with features
When you have different features, the best practice is to create
separate files for them. You used the Feature tag several times. In
reality, you should just have one Feature tag and then use the scenario
tag for that. Each scenario should come with a description as to what
the scenario is.