Removed shallow step definitions
The step definitions file was there to prevent the pipeline from
failing. The file was not clean at  all and everything was clutter in
there. What we really want is to implement the steps as part of the
development of new features and organize them in a way that makes sense
from a maintainability point of view.